+++ tool/src/main/kotlin/com/lightrss/reader/RssScreens.kt
@@ -869,7 +869,7 @@ class SettingsScreen(
                             modifier = Modifier.padding(top = 0.25f.gridUnitsAsDp()),
                         )
                         LightText(
-                            text = "VERSION 1.8.0",
+                            text = "VERSION 1.9.0",
                             variant = LightTextVariant.Superfine,
                             lighten = true,
                             modifier = Modifier.padding(top = 1f.gridUnitsAsDp()),